+++ src/client/js/app.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 // import SCSS
 import '../styles/footer.scss';
 import '../styles/form.scss';
@@ -7,19 +8,19 @@ import '../styles/container.scss';
 import '../styles/loader.scss';
 
 // import JS
-import {handleSubmit} from './handleSubmit';
-import {handleTab} from './handleTab';
-import {updateImage} from './updateImage';
-import {handleDate} from './handleDate';
-import {updateUI} from './updateUI';
-import {updateWeather} from './updateWeather';
-import {handleLoader} from './handleLoader';
-import {loopImage} from './loopImage';
-import {postData} from './postData';
-import {deleteTrip} from './deleteTrip';
+import { handleSubmit } from './handleSubmit';
+import { handleTab } from './handleTab';
+import { updateImage } from './updateImage';
+import { handleDate } from './handleDate';
+import { updateUI } from './updateUI';
+import { updateWeather } from './updateWeather';
+import { handleLoader } from './handleLoader';
+import { loopImage } from './loopImage';
+import { postData } from './postData';
+import { deleteTrip } from './deleteTrip';
 
 // import html
-import html from "../html/index.html";
+import html from '../html/index.html';
 
 // import media
 import github from '../media/github.png';
@@ -30,26 +31,26 @@ import compass from '../media/compass.png';
 
 // global export
 export {
-    handleSubmit,
-    handleTab,
-    updateImage,
-    handleDate,
-    updateUI,
-    updateWeather,
-    handleLoader,
-    loopImage,
-    postData,
-    deleteTrip
+  handleSubmit,
+  handleTab,
+  updateImage,
+  handleDate,
+  updateUI,
+  updateWeather,
+  handleLoader,
+  loopImage,
+  postData,
+  deleteTrip
 }
 
 // Run handleDate function when page loaded
-window.addEventListener("load", (event) => {
-    Client.handleDate()
+window.addEventListener('load', (event) => {
+  Client.handleDate()
 });
 
 // Automatic change background image with setTimeout loop
-window.addEventListener("submit", (event) => {
-    event.preventDefault();
+window.addEventListener('submit', (event) => {
+  event.preventDefault();
 
-    Client.loopImage();
+  Client.loopImage();
 });
